BackEndHub/requirements.txt
uvicorn>=0.20.0
fastapi>=0.89.1
pydantic>=1.8.0,<2.0.0
PyYAML>=6.0
sse-starlette>=1.8.2
asyncio>=3.4.3
shapely>=2.0.2
mysql-connector-python>=8.3.0
ConfigParser>=6.0.0 